+++ catalog/app/containers/Bucket/Overview/gql/BucketConfig.graphql
