+++ catalog/app/containers/Bucket/Overview/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Overview'
